requirements.txt
# ETL QORE - Dependencias
# ========================

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Selenium (Automacao)
selenium>=4.0.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Database
pyodbc>=5.0.0           # Access
psycopg2-binary>=2.9.9  # PostgreSQL

# Calendar/Date
bizdays>=1.0.0

# Windows Integration
pywin32>=306

# Utilities
python-dotenv>=1.0.0

# Dash (Dashboard - opcional)
dash>=2.14.0
plotly>=5.18.0

# XML Processing (stdlib, nao requer install)
# xml.etree.ElementTree
